Image colorization

+++ guided_diffusion/vamp_models.py
@@ -193,7 +193,7 @@ class VAMP:
         ################
 
         if t[0] > 200:
-            eta_2 = 1 / (self.scale_factor[used_t[0]] * true_noise_var.sqrt().unsqueeze(0).repeat(r_2.shape[0], self.Q))
+            eta_2 = 1 / (self.scale_factor[used_t[0]] * true_noise_var.sqrt().unsqueeze(0).repeat(r_2.shape[0], self.Q)).float()
         else:
             tr = self.denoiser_tr_approx(new_r_2, gamma_2, mu_2, noise_var, noise)
             eta_2 = 1 / tr
